requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib
bcrypt==4.0.1
python-multipart
alembic
pytest
httpx